Tokens

+++ server/src/routes/accounts.ts
@@ -3,15 +3,35 @@ import { Request } from "../utils/Request";
 import { AccountService } from "../services/AccountService";
 import {
     AccountRecord,
+    LoginRequest,
     RegisterRequest,
 } from "../../../shared/resource_models/account";
 import { HTTP, sendError, validateRequest } from "jack-hermanson-ts-utils";
-import { newAccountSchema } from "../models/Account";
+import { loginSchema, logoutSchema, newAccountSchema } from "../models/Account";
+import {
+    LogOutRequest,
+    TokenLoginRequest,
+    TokenRecord,
+} from "../../../shared/resource_models/token";
+import { tokenLoginSchema } from "../models/Token";
+import { auth } from "../middleware/auth";
+import { authorized } from "../utils/functions";
+import { Clearance } from "../../../shared/enums";
 
 export const router = express.Router();
 
 // get accounts
-router.get("/", async (req: Request<any>, res: Response) => {
+router.get("/", auth, async (req: Request<any>, res: Response) => {
+    if (
+        !authorized({
+            requestingAccount: req.account,
+            minClearance: Clearance.ADMIN,
+            res: res,
+        })
+    ) {
+        return;
+    }
+
     const accounts = await AccountService.getAccounts();
     res.json(accounts);
 });
@@ -42,8 +62,18 @@ router.post(
 
 router.delete(
     "/:id",
+    auth,
     async (req: Request<{ id: number }>, res: Response<boolean>) => {
         try {
+            if (
+                !authorized({
+                    requestingAccount: req.account,
+                    minClearance: Clearance.ADMIN,
+                    res: res,
+                })
+            ) {
+                return;
+            }
             const deleted = await AccountService.delete(req.params.id, res);
             if (!deleted) {
                 return;
@@ -52,5 +82,65 @@ router.delete(
         } catch (error) {
             sendError(error, res);
         }
+    }
+);
+
+router.post(
+    "/login",
+    async (req: Request<LoginRequest>, res: Response<TokenRecord>) => {
+        try {
+            if (!(await validateRequest(loginSchema, req, res))) {
+                return;
+            }
+            const loginRequest: LoginRequest = req.body;
+            const token = await AccountService.login(loginRequest, res);
+
+            if (!token) {
+                return;
+            }
+
+            res.json(token);
+        } catch (error) {
+            sendError(error, res);
+        }
+    }
+);
+
+router.post(
+    "/token",
+    async (req: Request<TokenLoginRequest>, res: Response<TokenRecord>) => {
+        try {
+            if (!(await validateRequest(tokenLoginSchema, req, res))) {
+                return;
+            }
+            const tokenLoginRequest: TokenLoginRequest = req.body;
+            const token = await AccountService.loginWithToken(
+                tokenLoginRequest,
+                res
+            );
+            if (!token) {
+                return;
+            }
+
+            res.json(token);
+        } catch (error) {
+            sendError(error, res);
+        }
+    }
+);
+
+router.post(
+    "/logout",
+    auth,
+    async (req: Request<LogOutRequest>, res: Response<boolean>) => {
+        if (!(await validateRequest(logoutSchema, req, res))) {
+            return;
+        }
+        const logOutRequest: LogOutRequest = req.body;
+        const success = await AccountService.logOut(logOutRequest, res);
+        if (!success) {
+            return;
+        }
+        res.send(success);
     }
 );
